Refactor VIF class object model
Currently the VIF class has a 'type' field that determines what
values are expected in the 'details' dictionary, which which
plugin is used to integration host actions.

There are ultimately many neutron mechanisms which map to the
same underlying libvirt VIF configuration, but have different
host integration work. For this reason, it is desirable to
disassociate the VIF configuration from the plugin actions.

This removes the 'type' field from the VIF class (along with
many other fields), and introduces subclasses to represent
the various different guest facing configuration approaches
for VIF. The goal is that the VIF subclass have enough info
to allow the guest XML (or non-libvirt equiv) to be created.

There is now an explicit 'plugin' field which provides
the name of the plugin class to use for host OS integration.

Thus, instead of there being a 1:1 mapping between Neutron
mechanisms and VIF types and plugin classes, there is now
an M:N mapping. ie, multiple different Neutron mechanisms
can all use the same VIF subclass to represent their config,
but provide different plugin classes to do their host OS
setup. This is key to avoiding the explosion in number of
VIF types.

NB, this has not updated any of the plugin impls, not
detailed how the existing "legacy" VIF types will be dealt
with. This set of VIF subclasses defines an idealized
minimal set of VIF types. When talking to old neutron
Nova will have to some how re-write the old VIF types to
use the new VIF subclasses as needed, and provide some
suitable plugin impls for them. For modern VIF types
the vendor will be responsible for providing the plugin
impls. This suggests, that the _plugins dir probably in
fact belongs in Nova.